Refactor model abstraction to class
